+++ libs/forms-react/src/index.ts
@@ -59,6 +59,7 @@ export type {
 } from './lib/field/use-mask-field'
 export { useMaskField } from './lib/field/use-mask-field'
 export { useResolvedFieldProps } from './lib/field/use-resolved-field-props'
+export { useActionFormErrors } from './lib/form/use-action-form-errors'
 export type {
   FormServerActionToaster,
   UseFormServerActionOptions,